refactor: sync delegations method (#260)
By removing the side effect many methods that read data in memory no
longer have to be async. I think this makes the API simplier.

I've added a _method_ that can be called to clean up expired
delegations, which shifts the responsibility onto the user.
Alternatively we could move the clean up into the methods that write
delegations? Open to other suggestions.

resolves #219