feat: add uploadCAR function (#329)
The PR adds an `uploadCAR` function to the client.

It differs from `Store.add` because it automatically shards the CAR
file, calls `Store.add` for each shard and finally registers an upload
(using `Upload.add`). This is very similar to `uploadFile` or
`uploadDirectory` except it skips the UnixFS encoding step.

Once merged, I'll propogate this to the
https://github.com/web3-storage/w3cli so that @lanzafame can upload BIG
CARs and have them automatically sharded.