feat: update access-api ucanto proxy to not need a signer (#390)
… and not sign proxyInvocation. using features coming in ucanto 4.2.0

Motivation:
* https://github.com/web3-storage/w3protocol/issues/325
* simplify access-api ucanto proxy using features added to ucanto in
https://github.com/web3-storage/ucanto/pull/199
* previously, the technique used to proxy the invocation was to issue a
new invocation (i.e. `proxyInvocation`) in the proxy server, and then
send that to the upstream. This had at least two limitations:
1. required the proxy server to be configured with a `options.signer` to
sign the `proxyInvocation`
2. for functional use in access-api and proxying upload-api, this proxy
`options.signer` also had to be configured pretty much identically to
the ucanto verifier with same did on the upstream, including requiring
both to have the same private key
  * now
* you don't need an `options.signer` at all! so you definitely don't
need one creating signatures with the same private key as the upstream

Steps
* [x] release ucanto 4.2.0
https://github.com/web3-storage/ucanto/pull/200
* [x] update this source branch package.json + pnpm locks to upgrade
ucanto to 4.2.0
* [x] ensure `tsc` + tests pass here