fix: account for edge cases in polling abort (#763)
My previous implementation
(https://github.com/web3-storage/w3up/pull/756) relied on canceling a
timeout to stop the polling process. This had a number of edge cases,
including one where the signal was aborted before first round of polling
happened, which is quite common since we do some network calls before we
even get to this code.

Stop worrying about canceling the timeout since we'll just stop polling
whenever we execute `poll` in this new implementation - generally this
relies on a piece of "persistent" state (in the sense that
`signal.aborted` will continue to be true) rather than the assumption
that the correct timeout will get canceled, so is quite a bit more
robust.

Also incorporate a reformulation of the polling code from @alanshaw  - thanks Alan!