feat!: merge `@web3-storage/access-api` into `@web3-storage/upload-api` (#790)
![image](https://github.com/web3-storage/w3up/assets/1113/2634d66f-6731-4e04-bd9a-c954c6535da1)

As part of the transition from D1 to Dynamo for delegation and provision
storage, move service providers from `@web3-storage/access-api` to
`@web3-storage/upload-api`.

This change decouples this entire repository from any particular storage
implementation and should make it easier for third parties to create
w3up-compatible storage services. We now expose 2 storage interfaces -
`DelegationsStorage` and `ProvisionsStorage` - and an email provider
interface (right now simply called `Email`) that implementors like our
very own `w3infra` need to implement. A prototype implementation of
these interfaces can be found in
https://github.com/web3-storage/w3infra/pull/194 - feedback on that PR
was the basis for some of the public interface changes in this PR.

I've also taken this opportunity to drop support for the voucher
protocol and stop using the `spaces` table which appears to only support
that. I've also dropped the `accounts` table since it didn't seem to be
adding anything - the one query we made to it returned the account DID
that we used to query the table. Once we have a clearer picture of what
information we want to store in an account we can bring that back if
that makes sense.