fix!: update aggregate spec in client and api (#824)
Updates client and api based on
https://github.com/web3-storage/specs/commit/63846e529a77d200af4a4a8bdab4c4ac7320c107

It also starts relying on the
https://github.com/web3-storage/data-segment/ lib for the tests by
generating commP of randomly generated CARs. In next iteration, tests
should also include commP of commPs builder (needs
https://github.com/web3-storage/data-segment/pull/8)

Client API for submitting offer was changed to receive piece instead of
computing it. Main reason for this is that there is a need to build it
outside while keeping track of sizes. It could also happen here and
explode if not, even though this is an aggregate offer call and not
really the builder. So I would love feedback on API surface preferences.
This will not be used by end users as will be abstracted within our
services. The API will still need to validate if given commP of commPs
is good for the given offer.

BREAKING CHANGE: aggregate capabilities now have different nb properties
and aggregate client api was simplified