fix: store add should validate size right away (#917)
`store/add` enforcing max single put size in context of w3filecoin
https://github.com/web3-storage/w3up/issues/827

At this moment, we only fail to provide the presigned URL for upload if
size is bigger, but we perform our DB insertions. storeTable insert, and
more recently the allocation

- The reasoning for this is described in code: `checking this last, as
larger CAR may already exist in bucket from pinning service fetch. we
only want to prevent this here so we don't give the user a PUT url they
can't use.`
- That makes sense to me from billing perspective. However, we will
allocate way more bytes to the space than realistically happens, as well
as show this in space store list.
- Regarding Pinning Service, we should consider chunking into smaller
CARs if that is realistic. Perhaps when copying from validation bucket
to destination
- 
I think that we should move up the code path this validation

---------

Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>