feat: compute piece CID in client (#925)
This PR alters the client to compute piece CIDs for each shard, and
exposes it via the `CARMetadata` provided to the `onShardStored`
callback.

I removed the `ShardStoringStream`. In hindsight, the only real utility
it was providing was parallelizing the calls to `Store.add(...)`. I
swapped in a parallel transform stream that allows us to do the same
thing, but also concurrently calculate the piece CID of the shard.

Note: the client does not yet do anything with the piece CID, but in
theory it will be trivial to collect them and then send a bunch of
invocations to trigger filecoin.