chore: cleanup space DID types and schema (#928)
@gozala and I are discussing the possibility of allowing `did:web`s to
be used as spaces, in order to enable us to `access/delegate`
capabilities on service DIDs. As I explored this possibility, I noticed
that we are currently a bit sloppy with space DID definitions.

This change cleans up and moves Space DID definition to the capabilities
package. It does not make any changes to the interfaces yet, but puts us
in a better place to make the change we are discussing, alongside just
making our code cleaner.