feat: revocation handler (#960)
Implements `ucan/revoke` capability and a ucanto server hook that
performs the revocation checks.

I ended up changing revocations store interface to avoid serializing /
deserializing UCN CIDs and to layout data in a way that it's would be
more efficient to perform necessary lookups.

☹️ I'm not super happy about the way we have to interface with ucanto
hook and probably it would make sense to change hook API so it's more
like `query` on the revocations store. If we do that we would be able to
uplift logic from `utils/revocation.js` into ucanto and simplify things
here. That said I it's probably better to land thing here first and do
the uplifting if we'll have time to do so.

---------

Co-authored-by: Alan Shaw <alan.shaw@protocol.ai>