fix: avoid duplicates on aggregator buffer concat (#1259)
While we know that some queues/streams we use are "at least 1 delivery",
we have barely seen any duplicate (see
https://filecoinproject.slack.com/archives/C02BZPRS9HP/p1704731857750279).

However, when we had a considerable amount of failures and retries,
looks like the "at least 1 delivery" primitive got more significant and
we actually got some repeated items. This tries to decrease likelihood
of repeated items. Of course that, if they actually make their way into
different aggregates this won't be effective, but same piece being added
in same aggregate should not happen anymore.

For this, we can just do the concat manually of both buffers to check if
a piece is already there