fix: upload client does not need cors mode on presigned url put (#1295)
Apparently some fetch implementations may not have "mode", which is the
case for Cloudflare Workers for instance. CF worker deploying
upload-client is now failing because this "mode" is set. You can also
read more about the general problem in
https://github.com/aws/aws-sdk-js-v3/issues/1165

In summary, "mode" may not be implemented in fetch, but the default
appears to be cors anyway so this could be deleted without loss of
functionality:
https://developer.mozilla.org/en-US/docs/Web/API/Request/mode#default_mode.

Related to https://github.com/web3-storage/w3up/issues/1273