fix: configure max pieces (#1566)
In lambda we only have 15 minutes max to execute. Without significant
changes to the way the library works there is only so many hashes we can
perform.

This PR adds a new configuration parameter `maxAggregatePieces` which
serves to restrict the maximum number of hashes the aggregator will have
to perform to build an aggregate. This will be tuned to be as high as
possible to allow execution within the time bounds. The PR also adds
logging in order to determine this limit.