fix: Find the receipts URL from the connection URL (#1565)
We were defaulting to the production URL unless a `receiptsEndpoint` was
given in the options, and then never providing one. Passing it along
would probably be the better solution, but there are a lot of places to
get that wrong as things stand right now. It's not clear to me where the
receipt URL really should live--eg., whether it should properly be part
of the connection config.