feat: external login (#1629)
refs https://github.com/storacha/RFC/pull/43

This PR adds a new method `externalLogin()` that allow login to the
system via an external method (e.g. OAuth). It essentially bypasses
email confirmation, assuming this is done out of band. We re-join the
regular auth flow at the point when we claim the delegation (and
attestation) from an email to the agent.